Watch It Follows (2014)
We found the ending to be unsatisfying, but highly committed to realism.
The setting and period of the film seemed intentionally ambiguous,
mixing 80s vibes with a timeless sense of apparel and language. There
were fewer jump scares to make room for more anxiety-based terror. Some
takeaways we had about the metaphorical significance of the monster were
the haunting aspect of your sexual history and the links between past
partners.