Watch Babel (2006)
Bad things happened, but there were no bad people.
The plot elements that would connect the four different worlds were
complicated enough that we could not accurately predict them. Every
campy twist we forsaw was sidestepped in favor for a more realistic and
less dramatic resolution. The film exists within a glimpse of a greater
timeline, leaving us to fill in the space before and after the film
with intimate hints but a low volume of content. The non-linear
timelines made for an interesting unraveling of the stream of events
at the ending of the film.