Watch Don't Breathe (2016)
A classic horror/thriller plot with an atypical villain. Some of the
camera work we most appreciated was the 'Vertigo shot' in the air vent
and the editing during the basement night vision scene. The Evil Dead
team's influence was palpable with content like the kidnapped girl, the
insemination scene, the grab right after kicking out the grate, and the
bone crunching fall down the air vent. Don't Breathe brings close-
quartered tension to the screen in place of the grotesque squeamishness
or cheaper jump scares. "Artfully Terrifying."