Watch Under the Silver Lake (2018)
We found the film to be incredibly abstract and overwhelmingly symbolic,
so much so that many elements and scenes were never revisited and were
seemingly unlinked to the progression of the greater plot. For example,
the shadow figure that follows the main character through the woods and
the transiently recurring pirate character are never given further
context. This erraticism is likely due to the allegedly high number of
allusions to other movies, most of which we were unable to identify. We
also noted strong elements of foreshadowing (e.g. Jesus and the Three
Brides of Dracula predicting the structure of the cult). The style
choices (settings, clothing, etc.) were reflective of present-day youth
culture (read hipsters), while maintaining a feeling of homage to
classic thriller/horror/mystery films. Although the plot was
interestingly convoluted, we noted a general lack of cinematographic
innovation or particularly dazzling acting performances.